% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialproportions.R
\name{BinomialProportions}
\alias{BinomialProportions}
\title{Bayesian Binomial Proportions test}
\usage{
BinomialProportions(x, n)
}
\arguments{
\item{x}{Number of successes.}

\item{n}{A vector of the number of trials.}
}
\value{
A vector of the probability of each alternative being best.
}
\description{
Fits a Bayesian binomial proportions model for a Bernoulli experment, 
returning the probability of each alterntive being best.
}
