% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomialproportions.R
\name{MultinomialProportions}
\alias{MultinomialProportions}
\title{Bayesian Multinomial Proportions test}
\usage{
MultinomialProportions(observed, expected = rep(mean(observed),
  length(observed)))
}
\arguments{
\item{observed}{Vector of the number of observed choices.}

\item{expected}{Vector of expected choices. If missing, it is rep(mean(observed), length(observed))}
}
\value{
A vector of the probability of each alternative being best.
}
\description{
Fits a Bayesian MultinomialProportions proportions model for a multinomial experment, 
returning the probability of each alterntive being best.
}
